requirements.txt
fastapi
requests
jinja2
azure-cognitiveservices-vision-computervision
msrest
python-dotenv
pillow
pyzbar
python-multipart
pyodbc